Fix CI build failure: cast maxDimension to uint for MagickGeometry
MagickGeometry's constructor takes uint, uint. The literal 2000 used
elsewhere in this codebase converts implicitly as a compile-time
constant, but the new parameterized maxDimension (used for the
thumbnail's 500px resize) is a runtime int and needs an explicit cast.

CS1503 caught by the deploy workflow's dotnet publish step, which is
the first real compile of this PR (no .NET SDK was available in the
authoring sandbox to catch it earlier).

+++ RecipeHelper/Services/StorageService.cs
@@ -171,7 +171,7 @@ namespace RecipeHelper.Services
 
                 if (image.Width > maxDimension || image.Height > maxDimension)
                 {
-                    image.Resize(new MagickGeometry(maxDimension, maxDimension)
+                    image.Resize(new MagickGeometry((uint)maxDimension, (uint)maxDimension)
                     {
                         IgnoreAspectRatio = false
                     });
